backend/requirements.txt
Flask
openai
gTTS
python-dotenv
better_profanity
